Improve test file practices
By awaiting async userEvents and awaiting screen.findByText
for transitions, I was able to remove the artificial
flushPromises() utility (and the testing folder,
since that's all that was in there).

Also I switched from individual mock clearing to jest.clearAllMocks()
to set a pattern that's less prone to state leak.